+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `devctl gen precommit`: Generates opinionated `.pre-commit-config.yaml` file based on projects language,
+  flavours and name
+
 ## [7.36.0] - 2026-03-06
 
 ### Added